pyproject.toml
[project]
name = "graphrag-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langchain-neo4j>=0.9.0",
    "langchain-openai>=1.2.1",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
]
